+++ src/types/spelling.d.ts
@@ -1,0 +1,2 @@
+declare module 'spelling';
+declare module 'spelling/dictionaries/en_US';