+++ CHANGELOG.md
@@ -4,12 +4,13 @@ All notable changes to the "str-conv" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+## [1.0.4]
 
 - Fix url match bug.
 - Fix a selection unit test failure.
 - Fix english word match to base64 bug.
 - Fix hover special char `)` render bug.
+- Support C/C++ string literal parsing.
 
 ## [1.0.3]